Grayboxed necessities in Unity
Every model is inside of an empty game object. When programming, call upon the empty parent objects, not the models inside, that way the models can be replaced with minimal effort.

X button needs to be programmed to close program

GameOver UI needs to be programmed to update HighScore and Score

Score needs to be updated in UI

Tags need to be added to Bumpers vs Targets, not gonna do that until y'all decide point values and if all bumpers should be worth the same points